Renamed GeoTrellis Constants to just Constants
